requirements.txt
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-ollama==0.2.2
pydantic==2.10.3
chromadb==0.4.22
sentence-transformers==2.3.1
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.27.0
python-multipart==0.0.6
jinja2==3.1.3
pandas==2.1.4
plotly==5.18.0
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
aiofiles==23.2.1